feat: #253 카카오톡 공유하기 기능 모듈화
- public/javascripts/totalShare.js로 분리.
- 매개변수 통일을 위해 article.js에 객체 생성.
