+++ saver-web/public/javascripts/infiniteScroll.js
@@ -5,42 +5,7 @@ var page = INITIAL_PAGE;
 var isUsed = false;
 var articleList = document.querySelector('.article-list');
 
-const KaKaoShare = (id, article) => {
-  if (!Kakao.isInitialized()) Kakao.init('카카오톡 JS 키');
-  const targetLink = `${document.location.href}articles/detail/${article.id}`;
-
-  Kakao.Link.createDefaultButton({
-    container: `#${id}`,
-    objectType: 'feed',
-    content: {
-      title: article.headline,
-      description: unescape(article.briefing),
-      imageUrl: article.image,
-      link: {
-        webUrl: targetLink,
-        mobileWebUrl: targetLink,
-        androidExecutionParams: 'test',
-      },
-    },
-    buttons: [
-      {
-        title: '웹으로 이동',
-        link: {
-          webUrl: targetLink,
-          mobileWebUrl: targetLink,
-        },
-      },
-      {
-        title: '앱으로 이동',
-        link: {
-          webUrl: targetLink,
-          mobileWebUrl: targetLink,
-          androidExecutionParams: `articles/detail/${article.id}`,
-        },
-      },
-    ],
-  });
-};
+import KaKaoShare from './totalShare.js';
 
 const getPage = () => {
   $.ajax({
@@ -51,11 +16,9 @@ const getPage = () => {
       articles.map(function (article) {
         articleList.insertAdjacentHTML('beforeend', makeTemplate(article));
         const item = document.getElementById(`share-button_${article.id}`);
+        article.path = `articles/detail/${article.id}`
         if (item) KaKaoShare(item.id, article);
       });
-
-      // Array.from(articles).forEach((article) => {
-      // });
 
       isUsed = false;
       page += ADD_PAGE;
