+++ saver-web/views/user/article.ejs
@@ -40,3 +40,4 @@
 
 <script defer src="https://developers.kakao.com/sdk/js/kakao.min.js"></script>
 <script type="module" src="/javascripts/user/article.js"></script>
+<script defer type="module" src="/javascripts/totalShare.js"></script>
