+++ saver-web/views/user/index.ejs
@@ -27,3 +27,4 @@
 <script defer src="https://developers.kakao.com/sdk/js/kakao.min.js"></script>
 <script defer type="module" src="/javascripts/user/index.js"></script>
 <script defer type="module" src="/javascripts/infiniteScroll.js"></script>
+<script defer type="module" src="/javascripts/totalShare.js"></script>
